New City.search_names setter pre save signal reciever.
You can see how to use it in admin.py. It contains city, regions,
countries, alternate names, and it's easy to use:

    from cities_light.models import City, to_search

    City.objects.filter(search_names__icontains=to_search(q))

Where 'q' may be anything you want. You can test it in the admin,
because there is a demonstration of how it works in the admin. This is
useful because it also shows what the autocomplete will suggest.

It behaves like google maps autocomplete, so that's as powerful as it
could get.